requirements.txt
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.4
Pillow==10.0.1
